Fix some compile error
