+++ paddle/cinn/hlir/pass/general_fusion_merge_pass/CMakeLists.txt
@@ -2,7 +2,7 @@ set(fusion_pass_file
     ${PADDLE_BINARY_DIR}/paddle/cinn/hlir/pass/use_general_pass.h)
 file(
   WRITE ${fusion_pass_file}
-  "#include \"paddle/cinn/common/macros.h\" // Generated by the paddle/cinn/hlir/pass/CMakeLists.txt.  DO NOT EDIT!\n\n"
+  "#include \"paddle/cinn/common/macros.h\" // Generated by the paddle/cinn/hlir/pass/CMakeLists.txt. DO NOT EDIT!\n\n#pragma once\n\n"
 )
 
 function(find_fusion_pass_register FILENAME ADD_PATH PATTERN)
@@ -19,7 +19,7 @@ function(find_fusion_pass_register FILENAME ADD_PATH PATTERN)
   endif()
 endfunction()
 
-file(GLOB_RECURSE FUSION_PASS_HEADER ./*.h)
+file(GLOB_RECURSE FUSION_PASS_HEADER ./*.cc)
 foreach(file ${FUSION_PASS_HEADER})
   find_fusion_pass_register("${file}" ${fusion_pass_file}
                             "CINN_REGISTER_FUSION_PASS")
@@ -27,4 +27,6 @@ endforeach()
 
 core_gather_headers()
 
-gather_srcs(cinnapi_src SRCS graph_group_fuse_helper.cc)
+gather_srcs(
+  cinnapi_src SRCS default_horizontal_fuse_pass.cc default_input_fuse_pass.cc
+  default_vertical_fuse_pass.cc default_recompute_fuse_pass.cc)
