+++ paddle/cinn/hlir/pass/general_fusion_merge_pass/graph_group_input_fuse_pass_ctx.h
@@ -21,6 +21,8 @@ namespace cinn {
 namespace hlir {
 namespace pass {
 
+class GraphGroupInputFusePassCtx;
+
 class GraphGroupInputFusePassCtx final : public InputFusePassCtx {
  public:
   GraphGroupInputFusePassCtx(
