+++ paddle/cinn/hlir/pass/general_fusion_merge_pass/graph_group_lightware_fuse_pass_ctx.h
@@ -21,6 +21,8 @@ namespace cinn {
 namespace hlir {
 namespace pass {
 
+class GraphGroupLightwareFusePassCtx;
+
 class GraphGroupLightwareFusePassCtx final : public LightwareFusePassCtx {
  public:
   GraphGroupLightwareFusePassCtx(
