+++ paddle/cinn/hlir/pass/general_fusion_merge_pass/vertical_fuse_util.h
@@ -14,64 +14,24 @@
 
 #pragma once
 
-#include "paddle/cinn/hlir/pass/general_fusion_merge_pass/fusion_pass_registrar.h"
+#include "paddle/cinn/api/op_group.h"
+#include "paddle/cinn/hlir/framework/op.h"
 #include "paddle/cinn/hlir/pass/general_fusion_merge_pass/lightware_fuse_pass_ctx.h"
-#include "paddle/cinn/hlir/pass/general_fusion_merge_pass/vertical_fuse_pass.h"
 #include "paddle/cinn/hlir/pass/general_fusion_merge_pass_utils.h"
 
 namespace cinn {
 namespace hlir {
 namespace pass {
 
-class DefaultVerticalFusePass final : public VerticalFusePass {
- public:
-  DefaultVerticalFusePass() : VerticalFusePass() {}
+using OpGroupPtr = api::OpGroup;
+using framework::OpPatternKind;
 
-  int Benefit() const override { return 100; }
-
-  void operator()(LightwareFusePassCtx* ctx) const override {
-    const auto& producer = ctx->PickOpGroup();
-    const OpGroupList consumers = [&]() {
-      OpGroupList consumers;
-      for (const auto& consumer : producer.consumers()) {
-        consumers.push_back(consumer);
-      }
-      return consumers;
-    }();
-    if (consumers.size() == 0) {
-      return;
-    }
-
-    std::vector<OpGroupPtr> candidates;
-    for (int i = 0; i < consumers.size(); ++i) {
-      const auto& consumer = consumers.at(i);
-      if (!DetectFusabilityByKind(ctx, producer, consumer)) {
-        break;
-      }
-      candidates.push_back(consumer);
-    }
-    if (candidates.size() == consumers.size() &&
-        producer.kind() == framework::kElementWise) {
-      return;
-    }
-
-    for (int i = 0; i < consumers.size(); ++i) {
-      const auto& consumer = consumers.at(i);
-      if (!DetectFusabilityByKind(ctx, producer, consumer)) {
-        continue;
-      }
-      if (ctx->fuse_helper().DetectCycleIfFuse(producer, consumer)) {
-        VLOG(4) << "Can't fuse because detect cycle";
-        continue;
-      }
-      ctx->MarkFusible(producer, consumer);
-    }
-  }
-
+struct VerticalFuseUtil {
   using KindKeyT = std::pair<OpPatternKind, OpPatternKind>;
-  bool DetectFusabilityByKind(LightwareFusePassCtx* ctx,
-                              const OpGroupPtr& src,
-                              const OpGroupPtr& dst) const {
+
+  static bool DetectFusabilityByKind(LightwareFusePassCtx* ctx,
+                                     const OpGroupPtr& src,
+                                     const OpGroupPtr& dst) {
     const KindKeyT kind_pair(src.kind(), dst.kind());
     const auto& map = GetConditionMap();
     const auto& iter = map.find(kind_pair);
@@ -92,41 +52,35 @@ class DefaultVerticalFusePass final : public VerticalFusePass {
 
   static std::map<KindKeyT, ConditionT> RawConditionMap() {
     return std::map<KindKeyT, ConditionT>{
-        {{OpPatternKind::kElementWise, framework::kElementWise},
-         &DefaultVerticalFusePass::IsSameSize},
+        {{OpPatternKind::kElementWise, framework::kElementWise}, &IsSameSize},
         {{OpPatternKind::kElementWise, framework::kBroadcast},
-         &DefaultVerticalFusePass::ElementwiseFuseBroadcast},
+         &ElementwiseFuseBroadcast},
         {{OpPatternKind::kElementWise, framework::kInjective},
-         &DefaultVerticalFusePass::HorizontalWithInjective},
+         &HorizontalWithInjective},
         {{OpPatternKind::kElementWise, framework::kReduction},
-         &DefaultVerticalFusePass::ElementwiseFuseReduce},
+         &ElementwiseFuseReduce},
 
-        {{OpPatternKind::kBroadcast, framework::kElementWise},
-         &DefaultVerticalFusePass::IsSameSize},
-        {{OpPatternKind::kBroadcast, framework::kBroadcast},
-         &DefaultVerticalFusePass::IsSameSize},
+        {{OpPatternKind::kBroadcast, framework::kElementWise}, &IsSameSize},
+        {{OpPatternKind::kBroadcast, framework::kBroadcast}, &IsSameSize},
         {{OpPatternKind::kBroadcast, framework::kInjective},
-         &DefaultVerticalFusePass::HorizontalWithInjective},
+         &HorizontalWithInjective},
         {{OpPatternKind::kBroadcast, framework::kReduction},
-         &DefaultVerticalFusePass::BroadcastFuseReduce},
+         &BroadcastFuseReduce},
 
-        {{OpPatternKind::kInjective, framework::kElementWise},
-         &DefaultVerticalFusePass::IsSameSize},
-        {{OpPatternKind::kInjective, framework::kBroadcast},
-         &DefaultVerticalFusePass::IsSameSize},
+        {{OpPatternKind::kInjective, framework::kElementWise}, &IsSameSize},
+        {{OpPatternKind::kInjective, framework::kBroadcast}, &IsSameSize},
         {{OpPatternKind::kInjective, framework::kInjective},
-         &DefaultVerticalFusePass::HorizontalWithInjective},
+         &HorizontalWithInjective},
         {{OpPatternKind::kInjective, framework::kReduction},
-         &DefaultVerticalFusePass::InjectiveHorizontalWithReduce},
+         &InjectiveHorizontalWithReduce},
 
         {{OpPatternKind::kReduction, framework::kElementWise},
-         &DefaultVerticalFusePass::ReduceFuseElementwise},
+         &ReduceFuseElementwise},
         {{OpPatternKind::kReduction, framework::kBroadcast},
-         &DefaultVerticalFusePass::ReduceFuseBroadcast},
+         &ReduceFuseBroadcast},
         {{OpPatternKind::kReduction, framework::kInjective},
-         &DefaultVerticalFusePass::HorizontalWithInjective},
-        {{OpPatternKind::kReduction, framework::kReduction},
-         &DefaultVerticalFusePass::ReduceFuseReduce},
+         &HorizontalWithInjective},
+        {{OpPatternKind::kReduction, framework::kReduction}, &ReduceFuseReduce},
     };
   }
 
@@ -188,6 +142,3 @@ class DefaultVerticalFusePass final : public VerticalFusePass {
 }  // namespace pass
 }  // namespace hlir
 }  // namespace cinn
-
-CINN_REGISTER_FUSION_PASS(DefaultVerticalFusePass,
-                          cinn::hlir::pass::DefaultVerticalFusePass);
