+++ paddle/cinn/hlir/pass/use_pass.h
@@ -15,6 +15,7 @@
 #pragma once
 
 #include "paddle/cinn/common/macros.h"
+#include "paddle/cinn/hlir/pass/use_general_pass.h"
 
 CINN_USE_REGISTER(InferShape)
 CINN_USE_REGISTER(OpFusion)
